@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 2.8.3)
 project(pips)
 
-#LIST(APPEND CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake_modules)
+add_compile_options(-Wall -std=c++14)
 
 
 ## Find catkin macros and libraries
