@@ -2,6 +2,9 @@
 #define PIPS_IMAGE_GEOMETRY_CONVERTER_H
 
 #include <pips/collision_testing/geometry_models/geometry_models.h>
+#include <pips/collision_testing/geometry_models/cylinder.h>
+#include <pips/collision_testing/geometry_models/box.h>
+
 #include <pips/collision_testing/image_geometry_models/cylinder.h>
 #include <pips/collision_testing/image_geometry_models/box.h>
 #include <memory>
@@ -32,7 +35,8 @@ namespace pips
               ptr = std::make_shared<pips::collision_testing::image_geometry_models::Cylinder>(std::static_pointer_cast<const pips::collision_testing::geometry_models::Cylinder>(model), model_pose);
               break;
             case 1:
-              //ptr = std::make_shared<pips::collision_testing::image_geometry_models::Box>(std::static_pointer_cast<const pips::collision_testing::geometry_models::Box>(model), model_pose);
+              // ROS_WARN_STREAM("WARNING! Box model not currently working with ImageGeometryConverter!" );
+              ptr = std::make_shared<pips::collision_testing::image_geometry_models::Box>(std::static_pointer_cast<const pips::collision_testing::geometry_models::Box>(model), model_pose);
               break;
             default:
               ROS_WARN_STREAM("WARNING! Attempted to convert a geometry model (" << model->type_id_ << ") for which no conversion is known!");
